requirements.txt
proxmoxer
requests
python-dotenv
mcp>=1.0.0
uvicorn[standard]